requirements.txt
streamlit
pandas
requests
faiss-cpu
sentence-transformers
groq